requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.7
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==3.2.23
django-celery-results==2.5.1
django-simple-history==3.3.0
djangorestframework==3.13.1
importlib-metadata==4.13.0
kombu==5.2.4
prompt-toolkit==3.0.42
pytz==2023.3.post1
redis==5.0.1
sqlparse==0.4.4
typing_extensions==4.7.1
vine==5.1.0
wcwidth==0.2.12
zipp==3.15.0
